Start work on 0.3 based on Vaadin 7.6.0
TableSelection is now a SelectionModel instead of an Extension
