+++ GridExtensionPack-demo/src/main/java/org/vaadin/teemusa/gridextensions/demo/DemoContentLayout.java
@@ -10,7 +10,7 @@ import org.vaadin.teemusa.gridextensions.pagedcontainer.PageChangeEvent;
 import org.vaadin.teemusa.gridextensions.pagedcontainer.PageChangeListener;
 import org.vaadin.teemusa.gridextensions.pagedcontainer.PagedContainer;
 import org.vaadin.teemusa.gridextensions.pagedcontainer.PagedContainer.PagingControls;
-import org.vaadin.teemusa.gridextensions.tableselection.TableSelectionExtension;
+import org.vaadin.teemusa.gridextensions.tableselection.TableSelectionModel;
 
 import com.vaadin.data.Container.Indexed;
 import com.vaadin.data.Item;
@@ -30,156 +30,149 @@ import com.vaadin.ui.VerticalLayout;
 
 public class DemoContentLayout extends VerticalLayout {
 
-    public DemoContentLayout() {
+	public DemoContentLayout() {
 
-        // Set up the PagedGridContainer with a backing Container.Indexed
-        final PagedContainer container = new PagedContainer(createContainer());
+		// Set up the PagedGridContainer with a backing Container.Indexed
+		final PagedContainer container = new PagedContainer(createContainer());
 
-        final Grid grid = new Grid(container);
+		final Grid grid = new Grid(container);
 
-        final PagingControls controls = container.setGrid(grid);
-        controls.setPageLength(5);
-        // Show page number 3 initially; Pages are 0-based indices
-        controls.setPage(3);
+		final PagingControls controls = container.setGrid(grid);
+		controls.setPageLength(5);
+		// Show page number 3 initially; Pages are 0-based indices
+		controls.setPage(3);
 
-        // Show it in the middle of the screen
-        setStyleName("demoContentLayout");
-        setSizeFull();
-        addComponent(grid);
-        setComponentAlignment(grid, Alignment.MIDDLE_CENTER);
+		// Show it in the middle of the screen
+		setStyleName("demoContentLayout");
+		setSizeFull();
+		addComponent(grid);
+		setComponentAlignment(grid, Alignment.MIDDLE_CENTER);
 
-        ContextClickExtension.extend(grid).addContextClickListener(
-                new ContextClickListener() {
+		ContextClickExtension.extend(grid).addContextClickListener(new ContextClickListener() {
 
-                    private GridContextMenu menu = new GridContextMenu(grid);
+			private GridContextMenu menu = new GridContextMenu(grid);
 
-                    @Override
-                    public void contextClick(ContextClickEvent event) {
-                        menu.setItemId(event.getItemId());
-                        menu.setColumn(event.getColumn());
-                        menu.open(event.getClientX(), event.getClientY());
-                    }
-                });
+			@Override
+			public void contextClick(ContextClickEvent event) {
+				menu.setItemId(event.getItemId());
+				menu.setColumn(event.getColumn());
+				menu.open(event.getClientX(), event.getClientY());
+			}
+		});
 
-        final TableSelectionExtension tableSelect = TableSelectionExtension
-                .extend(grid);
-        tableSelect.setMode(TableSelectionMode.CTRL);
+		final TableSelectionModel tableSelect = new TableSelectionModel();
+		grid.setSelectionModel(tableSelect);
+		tableSelect.setMode(TableSelectionMode.CTRL);
 
-        HorizontalLayout tableSelectionControls = new HorizontalLayout();
-        tableSelectionControls.setCaption("Table Selection Controls");
+		HorizontalLayout tableSelectionControls = new HorizontalLayout();
+		tableSelectionControls.setCaption("Table Selection Controls");
 
-        // Controls for testing different TableSelectionModes
-        for (final TableSelectionMode t : TableSelectionMode.values()) {
-            tableSelectionControls.addComponent(new Button(t.toString(),
-                    new ClickListener() {
-                        @Override
-                        public void buttonClick(ClickEvent event) {
-                            tableSelect.setMode(t);
-                        }
-                    }));
-        }
+		// Controls for testing different TableSelectionModes
+		for (final TableSelectionMode t : TableSelectionMode.values()) {
+			tableSelectionControls.addComponent(new Button(t.toString(), new ClickListener() {
+				@Override
+				public void buttonClick(ClickEvent event) {
+					tableSelect.setMode(t);
+				}
+			}));
+		}
 
-        addComponent(tableSelectionControls);
-        setComponentAlignment(tableSelectionControls, Alignment.BOTTOM_CENTER);
+		addComponent(tableSelectionControls);
+		setComponentAlignment(tableSelectionControls, Alignment.BOTTOM_CENTER);
 
-        // Controls for paging container. Next/Prev buttons and a jump to page
-        // text field.
-        HorizontalLayout pageControls = new HorizontalLayout();
-        pageControls.addComponent(new Button("Previous page",
-                new ClickListener() {
-                    @Override
-                    public void buttonClick(ClickEvent event) {
-                        controls.previousPage();
-                    }
-                }));
+		// Controls for paging container. Next/Prev buttons and a jump to page
+		// text field.
+		HorizontalLayout pageControls = new HorizontalLayout();
+		pageControls.addComponent(new Button("Previous page", new ClickListener() {
+			@Override
+			public void buttonClick(ClickEvent event) {
+				controls.previousPage();
+			}
+		}));
 
-        // Text field only allows integer values
-        final TextField textField = new TextField("Page count: "
-                + controls.getPageCount());
-        textField.addValidator(new Validator() {
+		// Text field only allows integer values
+		final TextField textField = new TextField("Page count: " + controls.getPageCount());
+		textField.addValidator(new Validator() {
 
-            @Override
-            public void validate(Object value) throws InvalidValueException {
-                try {
-                    Integer.parseInt(value.toString());
-                } catch (NumberFormatException e) {
-                    throw new InvalidValueException(
-                            "Could not convert value to int");
-                }
-            }
-        });
+			@Override
+			public void validate(Object value) throws InvalidValueException {
+				try {
+					Integer.parseInt(value.toString());
+				} catch (NumberFormatException e) {
+					throw new InvalidValueException("Could not convert value to int");
+				}
+			}
+		});
 
-        // Starting value of 4 (page index 3)
-        textField.setValue("" + (controls.getPage() + 1));
-        textField.addValueChangeListener(new ValueChangeListener() {
+		// Starting value of 4 (page index 3)
+		textField.setValue("" + (controls.getPage() + 1));
+		textField.addValueChangeListener(new ValueChangeListener() {
 
-            @Override
-            public void valueChange(ValueChangeEvent event) {
-                controls.setPage(Integer.parseInt(event.getProperty()
-                        .getValue().toString()) - 1);
-            }
-        });
+			@Override
+			public void valueChange(ValueChangeEvent event) {
+				controls.setPage(Integer.parseInt(event.getProperty().getValue().toString()) - 1);
+			}
+		});
 
-        controls.addPageChangeListener(new PageChangeListener() {
+		controls.addPageChangeListener(new PageChangeListener() {
 
-            @Override
-            public void pageCountChange(PageChangeEvent event) {
-                textField.setCaption("Page count: " + controls.getPageCount());
-            }
+			@Override
+			public void pageCountChange(PageChangeEvent event) {
+				textField.setCaption("Page count: " + controls.getPageCount());
+			}
 
-            @Override
-            public void pageChange(PageChangeEvent event) {
-                textField.setValue("" + (controls.getPage() + 1));
+			@Override
+			public void pageChange(PageChangeEvent event) {
+				textField.setValue("" + (controls.getPage() + 1));
 
-            }
-        });
+			}
+		});
 
-        pageControls.addComponent(textField);
-        pageControls.addComponent(new Button("Next page", new ClickListener() {
-            @Override
-            public void buttonClick(ClickEvent event) {
-                controls.nextPage();
-            }
-        }));
-        pageControls.addComponent(new Button("Random page length",
-                new ClickListener() {
+		pageControls.addComponent(textField);
+		pageControls.addComponent(new Button("Next page", new ClickListener() {
+			@Override
+			public void buttonClick(ClickEvent event) {
+				controls.nextPage();
+			}
+		}));
+		pageControls.addComponent(new Button("Random page length", new ClickListener() {
 
-                    private Random r = new Random();
+			private Random r = new Random();
 
-                    @Override
-                    public void buttonClick(ClickEvent event) {
-                        controls.setPageLength(3 + r.nextInt(8));
-                    }
-                }));
+			@Override
+			public void buttonClick(ClickEvent event) {
+				controls.setPageLength(3 + r.nextInt(8));
+			}
+		}));
 
-        addComponent(pageControls);
-        setComponentAlignment(pageControls, Alignment.BOTTOM_CENTER);
+		addComponent(pageControls);
+		setComponentAlignment(pageControls, Alignment.BOTTOM_CENTER);
 
-        setMargin(true);
+		setMargin(true);
 
-        grid.setEditorEnabled(true);
-        for (Column c : grid.getColumns()) {
-            c.setHidable(true);
-        }
-    }
+		grid.setEditorEnabled(true);
+		for (Column c : grid.getColumns()) {
+			c.setHidable(true);
+		}
+	}
 
-    @SuppressWarnings("unchecked")
-    private Indexed createContainer() {
-        Indexed wrappedContainer = new IndexedContainer();
-        wrappedContainer.addContainerProperty("foo", String.class, "foo");
-        wrappedContainer.addContainerProperty("bar", Integer.class, 0);
-        // km contains double values from 0.0 to 2.0
-        wrappedContainer.addContainerProperty("km", Double.class, 0);
+	@SuppressWarnings("unchecked")
+	private Indexed createContainer() {
+		Indexed wrappedContainer = new IndexedContainer();
+		wrappedContainer.addContainerProperty("foo", String.class, "foo");
+		wrappedContainer.addContainerProperty("bar", Integer.class, 0);
+		// km contains double values from 0.0 to 2.0
+		wrappedContainer.addContainerProperty("km", Double.class, 0);
 
-        for (int i = 0; i <= 30; ++i) {
-            Object itemId = wrappedContainer.addItem();
-            Item item = wrappedContainer.getItem(itemId);
-            item.getItemProperty("foo").setValue("foo");
-            item.getItemProperty("bar").setValue(i);
-            item.getItemProperty("km").setValue(i / 5.0d);
-        }
+		for (int i = 0; i <= 30; ++i) {
+			Object itemId = wrappedContainer.addItem();
+			Item item = wrappedContainer.getItem(itemId);
+			item.getItemProperty("foo").setValue("foo");
+			item.getItemProperty("bar").setValue(i);
+			item.getItemProperty("km").setValue(i / 5.0d);
+		}
 
-        return wrappedContainer;
-    }
+		return wrappedContainer;
+	}
 
 }
