+++ GridExtensionPack-demo/src/main/java/org/vaadin/teemusa/gridextensions/demo/HeaderWrapExtensionLayout.java
@@ -1,15 +1,12 @@
 package org.vaadin.teemusa.gridextensions.demo;
 
 import java.util.Random;
-import com.vaadin.annotations.Theme;
-import com.vaadin.annotations.VaadinServletConfiguration;
+
 import org.vaadin.teemusa.gridextensions.wrappinggrid.WrappingGrid;
-import com.vaadin.server.VaadinRequest;
-import com.vaadin.server.VaadinServlet;
+
 import com.vaadin.ui.Button;
 import com.vaadin.ui.Button.ClickEvent;
 import com.vaadin.ui.Grid;
-import com.vaadin.ui.UI;
 import com.vaadin.ui.VerticalLayout;
 
 public class HeaderWrapExtensionLayout extends VerticalLayout {
