+++ GridExtensionPack/pom.xml
@@ -5,12 +5,12 @@
 	<groupId>org.vaadin.teemusa</groupId>
 	<artifactId>gridextensionpack</artifactId>
 	<packaging>jar</packaging>
-	<version>0.2.3</version>
+	<version>0.3-SNAPSHOT</version>
 	<name>GridExtensionPack Add-on</name>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<vaadin.version>7.5.10</vaadin.version>
+		<vaadin.version>7.6.0</vaadin.version>
 		<vaadin.plugin.version>${vaadin.version}</vaadin.plugin.version>
 
 		<!-- ZIP Manifest fields -->
