+++ pom.xml
@@ -5,7 +5,7 @@
 	<groupId>org.vaadin.teemusa</groupId>
 	<artifactId>gridextensionpack-root</artifactId>
 	<packaging>pom</packaging>
-	<version>0.2.3</version>
+	<version>0.3-SNAPSHOT</version>
 	<name>GridExtensionPack Add-on Root Project</name>
 
 	<modules>
